Improve traceback rendering of checkpoints
This commit improves the rendering of checkpoints in tracebacks.
Notably, it harmonizes all three variants:
 - `checkpoint()` is rendered as "checkpoint() -> at ..."
 - `checkpoint(String)` is rendered as "checkpoint -> foo"
 - `checkpoint(String, true)` is rendered as "checkpoint(foo) -> at ..."

Additionally, this commit ensures that the representation is similar for
Flux, Mono and ParallelFlux. It adds tests that validate both with and
without `onOperatorDebug()`.

Finally, this commit removes the "checkpoint" nature of tracebacks that
are added by the reactor-tool agent (via `MethodReturnSnapshot`).

Fixes #2819.