fix #2839 Buffer when downstream not ready
When downstream requests have been exhausted but items have been
buffered the current behaviour is to raise an error, this change instead
will leave the buffer in place in that circumstance and flush it at the
point of the next request. Complete signals that arrive while there is a
buffer waiting are now delayed until it is flushed.

Also fixed is the potential for negative request amounts being sent
upstream which is a trivial change to just not send those.