Automatically draft release notes on maintenance branches push
This commit introduces Release-Drafter in the publish.yml workflow,
for pushes categorized as "SNAPSHOT" (ie corresponding to the merging of
a PR).

The use of `filter-by-commitish` in the config and `commitish` (set to
the target branch) in the action config should ensure that drafts are
created for both the `main` branch and older maintenance branches.

Fix #1445.