Use Gradle's VersionCatalog for dependencies in a single place (#2853)
This PR enables the usage of Gradle 7 Version Catalog.

The TOML file is used, in a way that _should_ be compatible with RenovateBot.

The versions contain:
 - a set of `baseline` versions (replaces equivalents in gradle.properties)
 - direct versions which are used in several places

The remaining versions are directly defined inline in the libraries and plugins
sections.

In addition, the compatibility of the various syntaxes used for versions/version
references with Renovate Bot has been separately validated. Renovate now
detects the Kotlin version used and we pin it to 1.5.x updates only.