Fix draft release note generation for multiple branches/drafts (#2876)
This commit removes the (flawed) Release-Drafter configuration and
action, and replaces it with two actions:
 - one that generates the notes
 - one that creates and updates draft releases correctly

This is done because it appears Release-Drafter cannot reliably create
and update two separate drafts for two maintenance branches (it will
create one draft per branch, but will overwrite both drafts with the
data from the latest run).

This new workflow relies on a new custom Gradle task similar to
qualifyVersionGha which computes the last tag from which to create the
release note and the expected name of the draft release / tag from the
current (snapshot) version.

It also uses a SKIP value to ensure that non-snapshot pushes won't
trigger an update of the draft note.

As a side effect, the rendering of PR titles has been fixed to include the
links to PRs (which were previously stripped). The new action doesn't
support listing all contributors in one line, so they are mentioned on each
entry.

Fixes #1445.