Replace drafter workflow with github's simple generator config
This commit replaces the slightly overcomplicated release-drafter
workflow with a simple configuration that will let GitHub's own release
note generator categorize PRs.

GitHub generator doesn't allow to define an overall template for the
notes, nor does it allow customization of how each PR entry looks like.

That said, it is good enough in most cases and seems to deal correctly
with multiple release branches (when fetching and comparing against the
latest tag, it filters on the specified target branch for the release
draft).

The old workflow was more configurable, but also relying on more moving
pieces:
 - a specific Gradle task to compute the latest tag
 - the changelog generating action with non-default fromTag and toTag
 configuration
 - an action requiring a large scope (contents:write) just to be able to
 produce a drafted release

This solution is more limited, but also easier to reason about in terms
of security.