Prevent FluxBufferTimeout from making negative request
This commit puts a simple guard against negative request in the
BufferTimeoutSubscriber#requestMore method in order to cover cases
where discrepancies are introduced by the timeouts.

Request accounting is not perfect when both a size and a duration can
be specified, given that requests upstream are multiplicated in order
to get a chance at filling buffers.

Hence, this change puts a simple guard which ignores negative requests.

Fixes #2839.