[build] Action updates: fix pinning
This commit fixes the Renovate configuration for actions in order to use
SHA pinning. The config block is derived from the existing helper
(helpers:pinGithubActionDigests) but since we're configuring more stuff
like grouping and schedule, this commit uses an explicit approach.

It also changes the use of matchManagers to matchDepTypes (inspired by
the helper again).

Fixes #2900.