Run preliminary spotless/japicmp job in ci, fix japicmp of core (#2895)
This commit changes the `ci.yml` workflow so that it runs a first job to
check spotless and japicmp. The other tests, which run longer, depend on
that first job succeeding.

We output a few tips to fix spotless failures and to indicate japicmp
failures need to be discussed with the maintainers, as pointers intended
for contributors. These hints would be the first expanded thing one sees
when clicking on the failing check.

This commit also fixes the japicmp configuration in the reactor-core
module, which was broken and didn't fail the gradle build even if some
incompatibilities were reported in japi.txt.