[build] Action updates: fix pinning (#2901)
This commit fixes the Renovate configuration for actions in order to use
SHA pinning. The config block is derived from the existing helper
(helpers:pinGithubActionDigests) but since we're configuring more stuff
like grouping and schedule, this commit uses an explicit approach.

It also changes the use of `matchManagers` to `matchDepTypes` (again
inspired by the helper).

Fixes #2900.