Improve readability of preliminary how-to-fix hint in CI
The message is intended as a hint on how to fix failing steps above it.
This commit makes it more explicit, both in text and by the choice of
color coding.

Previously, the separate hints had some colored background, green and
orange respectively. This made it look like the japicmp step was the one
detected as failed, which was counterintuitive.

The message now explicitly tells the reader to look at the previous
steps in the job, and the colors have been changed to:
 - drive focus to the whole message by having a header with orange bg
 - highlight the individual calls to action with white font on black bg
 - highlight the gradle command with black font on white bg