Change behavior of switchOnNext/switchMap default 0 prefetch (#2956)
This commit changes the no-parameter variant of Flux switchMap and
switchOnNext operators to default to a behavior of prefetch = 0.

This was documented in 3.4.x when the parameterized variant was
deprecated. Note that the prefetch > 0 implementation is slated for
complete removal in 3.6.0.

Old prefetch behavior could be emulated with `limitRate` on inners.

Fixes #2607.