[build] Have jcstress part of slowerChecks
This commit makes jcstress tests part of the slowerChecks job
in the publish.yml workflow.

Latest runs of this workflow show 24min for prepare and 14min
for slowerChecks, so this should bring better equilibrium.

Additionally, we ensure that only reactor-core:test are run
(the only one with a "slow" tag), as the previous command
was running _all_ tests and only filtering the core ones.

Fixes #2805.