[agent] Avoid instrumenting methods with unwrappable returnType (#2973)
This commit adds a check to instrumentation of methods that return a
publisher type that cannot be wrapped while maintaining a compatible
type.

One such incompatibility occurs in `Mono#toProcessor()`. It was
instrumented because `MonoProcessor` is a `Mono` technically, which
led to a `MonoAssembly` wrapping type. But that is not compatible with
the exposed return type (`MonoProcessor`).

Fixes #2970.