Force white background on README marble diagrams
This allows better readability of the marble diagrams with a dark
background, which is useful since the README is often view on GitHub
(which offers a dark mode).

The canonical usage of the marble diagrams is the reference guide, which
has a white background style, so the SVG is left untouched.

Fixes #3004.