[doc] Increase width of refguide content for readability (#3009)
The reference guide #content div has `max-width: 62.5em`. This is a bit
small for large screens in landscape mode. So, this commit increases the
width.

For landscape, in FHD (1920x1080) window, ideally, `1920px` * `60%` +
`8em (div's padding)` = `80em` is a required width for readability. For
more wide window, the constant will be smaller than `60%` of the
window's width. Thus, `max-width` is desirable to be maximum of `80em`
and `60%` of a window.

For portrait (phone), the changes for landscape doesn't affect for
portrait because it changes `max-width` to be increase.

So, this commit set `max-width` to `max(80em, 60%)`.

Fixes #2784.