Add 'tap', a generic side-effect/observability operator (#3013)
This commit adds a new operator named `tap`, which allows to tap into signals
emitted by the previous step (onSubscribe, onNext, onComplete, onError) and to
the downstream Subscription (cancel, request) in a stateful manner.

A `SignalListenerFactory` can be provided in order to define and capture state,
both at assembly time and subscription time.

The factory is invoked at subscription time for each downstream `Subscriber`,
creating a dedicated `SignalListener` from the common assembly time state and
the context of the subscriber.

All of the signals are notified to these `SignalListener` which themselves can
maintain state between each signals, unlike individual doOnXxx operators.

The `SignalListener` is also notified of malformed signals, has handlers that
are called after termination (doAfterComplete, doAfterError) and at the very
beginning/end of the sequence (doFirst, doFinally).

Exceptions thrown from the `SignalListener` handlers will cause a termination
of the sequence (if possible), and no more handlers will be invoked. Instead,
the #handleListenerError uncaught hook is invoked.