Introduce new module reactor-core-micrometer (#3015)
The new module introduces an alternative more explicit way of bringing
metrics in reactor-core, both for Schedulers and for Flux and Mono. The
entry point is the `Micrometer` class.

The following metrics-related features in core are superseded:
 - global configuration previously done in core's `Metrics`
 - scheduler `ExecutorService` instrumentation previously done in
 `Scheduler#enableMetrics`
 - Flux/Mono `metrics()` operators replaced by `tap` + the factory from
  the new `Micrometer.metrics()` method

As a consequence, these metrics-related classes and methods in core are
deprecated, to be removed in 3.6.0 at the earliest. Note that most
logic needed to be duplicated or reproduced in the new module.

The new module lives in the core repository and shares the same groupId
as reactor-core (io.projectreactor). It is dedicated to Micrometer
instrumentation for reactor-core.

The new module brings an explicit dependency to Micrometer 1.10+, which
means that the old `Metrics.isInstrumentationAvailable` will always be
`true` when the module is on the classpath. Extra care should be taken
to avoid activating metrics via both the core features and their module
equivalents.