Scannable.tags() rework, add tagsDeduplicated() (#3037)
This commit reworks Scannable.tags() with a slight behavioral change
in the order in which duplicate tags are streamed, and adds a new
`tagsDeduplicated()` method to get the tags from the hierarchy of
parents without duplicates.

In `tags()`, for a tag key that is defined multiple times, the first value
to be iterated is now the one from the uppermost parent. That way, it
is easier to eg. iterate and push to a `Map` in order to get values that
are defined closer to the current Scannable step.

Implementation-wise, tag operators now use a List<Tuple2> to store
the Tuple2 tags since the Set<Tuple2> wasn't bringing any benefit.

Fixes #2542.