Remove Processors from public API (#3051)
This commit removes the Processor implementations from the public API.
These have had a Sinks alternative since 3.4.0, at which point they were
deprecated. In 3.4.19, the ability to subscribe a `Sinks.Many` to an
upstream publisher is partly restored in a non-deprecated fashion with
the `Sinks.ManyWithUpstream` specialization of `Sinks.Many`.

This commit makes the relevant processor implementations that are also
implementing the various flavor of Sinks package-private. Most have been
renamed to use the `SinkMany` prefix in class name for grouping in the
class list. Others like the `DelegateProcessor` have been removed
entirely.

This commit also amends the javadocs and reference guide to remove
mentions of processors (or tone them done) in the documentation.
Finally, japicmp exclusions have been added for removed classes.