Reduce MonoCollectList cancel() synchronized block
This commit reduces the length of the critical section in cancel() that
is synchronized. In particular, it removes upstream cancel() from said
critical section. This will prevent a deadlock in case the upstream
operator itself uses synchronized block between cancel() and methods
that call any of this subscriber's synchronized methods (eg. onNext).

Fixes #3052.