Add context-propagation-api dependency + ReactorContextAccessor (#3098)
This commit adds a dependency to `micrometer:context-propagation-api`,
in snapshot version. Temporarily, it switches switches the micrometer
BOM to the snapshot version as well (for compatibility with the above
new dependency).

Most importantly, this new dependency is used to implement a context
accessor, `ReactorContextAccessor`, to be picked up by ServiceLoader
mechanism (META-INF/services/ file also included).