Micrometer module: add observation(), deprecate global registry (#3104)
This commit introduces `Micrometer.observation(ObservationRegistry)`,
adds Context support to tap/`SignalListener` and deprecates a couple of
methods.

`.observation()` is a new tap-oriented listener similar to `.metrics()`
which uses the Micrometer Observation API under the cover. An explicit
registry is required to instantiate the listener.

The MicrometerObservationListener:
 - covers the entire lifecycle of a Flux/Mono subscription with an
 Observation and associated Scope
 - uses context-propagation-api to ensure the Scope hierarchy is
 correct

Compared to `.metrics()`, `.observation()` only concerns itself with
subscription and termination events (cancel, onComplete, onError).
It uses the `Observation#error(e)` mechanism to report onError cause.
Spans names exactly reflect the one defined by the `.name()` operator
(making use of the `Observation#contextualName` for that purpose).

One big change in this commit is the ability for the `tap` operator to
extract a redefined `Context` from its `SignalListener`, by calling
new method `addToContext` after having invoked `doFirst()`.

This gives the `.observation` an opportunity to update the context with
a Scope it just opened in `doFirst`.

The third part is the deprecation of `Micrometer#useRegistry` and
`Micrometer#getRegistry` methods, in order to enforce explicit passing
of a `MeterRegistry` (just like `ObservationRegistry` is required for
`.observation()`).

Note that `io.micrometer:context-propagation-api` is added as a direct
dependency of the module.