Partial reintroduction of Flux Processors as deprecated
This partially reverts #3051, by reintroducing Flux Processors
implementation as they were in 3.4.x (still deprecated).

This was done by copying the classes and making a few small adaptations
to DirectProcessor and SinkManyBestEffort (reintroducing the interface
DirectInnerContainer, as a package-private mean of mutualizing the
DirectInner code).

This soft-revert doesn't reintroduce the MonoProcessor, which was also
deleted as part of the above PR.