Remove Micrometer#useRegistry, remove notion of Clock (#3128)
This commit reworks the `reactor-core-micrometer` `Micrometer` API to
simplify things a bit and make the use of MeterRegistry more explicit:
 - remove `useRegistry` and the notion of a default global registry
 - remove `metrics()` and `metrics(MeterRegistry, Clock)`
 - replace the above by `metrics(MeterRegistry)`

The `Clock` parameter was a bit redundant, as it can always be provided
through the `MeterRegistry#config()`. For example:

```
new SimpleMeterRegistry(SimpleConfig.DEFAULT, myCustomClock);
```

This is a breaking change compared to previous milestones only.