Improve LoggerUtils support and add TestLoggerExtension (#3123)
This commit improves `LoggerUtils` and adds supporting JUnit5 extension
and annotations in reactor-core tests:

`LoggerUtils` now has support to _redirect_ rather than _copy/capture_
log messages when the logger factory is installed early.

In reactor-core tests, a `TestLoggerExtension` is added that sets up a
`TestLogger` and activates capture/redirection via `LoggerUtils`:
 - before the test, it creates a `TestLogger` (configured depending on
 test annotations) and sets up `LoggerUtils` with said `TestLogger`
 - as a `ParameterResolver` it injects the `TestLogger` into the test
 - after the test it `disableCapture()`

The Extension should be applied to individual tests. Convenience
annotations are provided that fine tune the `TestLogger` that will be
injected:
 - `@Capture` when the loggers should capture log output, ie. go both
 to original logger and TestLogger
 - `@Redirect` when the loggers should redirect log output, ie. only go
 to TestLogger

This commit further simplifies the `LoggerUtils` internals, putting more
state into the `CapturingFactory` rather than as a top level static
field.

Finally, it adds a ton of test coverage to both `TestLogger` and
`LoggerUtils`.