Add TimedScheduler to reactor-core-micrometer module (#3109)
This commit introduces a TimedScheduler wrapper to reactor-core-micrometer
module which wraps any Scheduler to add task metrics to it.

This _replaces_ Micrometer.enableSchedulerMetrics from previous milestones,
and is thus considered a breaking API change **vs milestones M1-M4**.

The class itself is package-private, with a factory method in Micrometer that
only present a `Scheduler` view of the `TimedScheduler`.

The deprecation notice in core's `Schedulers#enableMetrics()` has been
updated to reflect the new alternative.

An old test in core was named `TimedSchedulerTest`. This is confusing now
that there is an actual `TimedScheduler` so these tests have been melded
into `SingleSchedulerTest` instead.

This is a different approach from the core one which also Fixes #1201 again.