SourceException wrapper for main stream errors delivered to windows
In case of window operators, it might be beneficial to distinguish
between errors coming from the window chain of operators and errors
coming from the source sequence. `Exceptions.SourceException` has
been added and `Throwable`s coming from either the source
sequence, or, in case of window operators that accept `Publisher`s
to drive windowing, their errors also are wrapped.

This is a behavior change and if existing logic relied on the
properties of the delivered `Throwable` it will require reaching for
the `Throwable#getCause()` in case of `Exceptions.SourceException`.

An additional behavior change is in case of `FluxWindowWhen`,
which did not deliver source errors to the windows, despite
seemingly being intended by the implementation.

Resolves #2041