Document reactor-Core-Micrometer obs/meters/tags via enums (#3175)
This commit uses DocumentedMeter API to replace magic String constants
and make the meters and tags in Reactor-Core-Micrometer documented and
public.

The asciidoc generation now takes advantage of this to generate a
section in the reference guide, through the provided Micrometer tool.

Each of the 3 features (meter listener, observation listener and timed
Scheduler wrapper) get their own DocumentedMeter and one associated
doc generation task. We use a Copy task trick on the three resulting
asciidoc files to remove some of the hardcoded boilerplate text for a
better inclusion in the reference guide.

Relates to #3108.