Improve JapiCmp: avoid misses, improve reporting and exclusions (#3198)
This PR improves the japicmp integration in the build in several ways:

1) Add a finalizedBy task that prints a filtered report
The basic text report uses * and ! as markers for binary and source
incompatible changes respectively. Additionally, a MODIFIED class is
marked with ***. This task attempts at pinpointing the incompatible
changes only in the report, and printing these lines. This is called out
in the ci job when prepare step fails fast.

2) Ensure sourceModified changes are considered
See reactor/reactor#722.
If a change is binary compatible but not source compatible, using
`onlyBinaryCompatibleModified = true` will mask the issue. This
change uses `onlyModified = true` instead. This is slightly more
verbose, but this is alleviated by (1).

3) Bump to japicmp plugin v0.4.1, exclude NEW_DEFAULT_METHOD as a whole
The new `compatibilityChangeExcludes` parameter allows us to ignore all
occurrences of a default method added to an interface, instead of
having to exclude each such method one by one.