Micrometer Observation test: ensure assert fails the test
This commit ensures that in MicrometerObservationIntegrationTest, the
last assertion that checks the delay thread for leftover Span can
actually fail the test if an unexpected Span is discovered.

This is done by capturing the currentSpan() in the thread, but asserting
it in the main test body (otherwise the AssertionError would just get
logged).

Also adds a few comments as to why that particular assertion used to
fail for a while due to Brave defaults, and a javadoc comment on how to
run the test.

Fixes #3214.