Remove marker need for tap/handle context-propagation
This commit removes the notion of a `Context` marker inserted by the
`contextCapture()` operator as a precondition for `tap` and `handle` to
restore ThreadLocals from a `ContextSnapshot` implicitely.

Now, the operators only check that A) `context-propagation` library is
available at runtime and B) the downstream context is not empty.

The javadoc and reference guide have been amended to reflect that fact.

Fixes #3250.