Limit the number of Scheduler#disposeGracefully threads
This change modifies the `DisposeAwaiter` interface in order to be able
to poll for graceful shutdown termination status rather than await/block
for it.

It introduces a DisposeAwaiterRunnable with a small pool of threads
dedicated to polling the termination status after a graceful Scheduler
shutdown.

Previously, one Thread would be created for each Scheduler that is
disposed gracefully. While we don't expect this to be an issue in most
production applications, this can lead to hitting native thread limits
faster. Notably, stress tests around graceful disposal create a lot of
schedulers for that purpose.