Context propagation: use new APIs from latest snapshots (#3256)
This commit upgrades context-propagation to post-RC1 snapshots.
This allows for several polishes and improvements:

ContextSnapshot.setAllThreadLocalsFrom can be used as a way of
restoring thread locasl while avoiding intermediary representation.

With the introduction of removal methods in ContextRegistry, it becomes
more practical to use the ContextRegistry.getInstance global instance.
Following that, two of the three intermediate classes are no longer
needed and can be replaced with lambdas: ContextCaptureFunction and
ContextRestoreHandleConsumer.

Instead of calling the ContextPropagation methods with a subscriber
(which necessitates some mocking in tests), we now take a Supplier
of Context. A CoreSubscriber::currentContext method reference can
do the trick in production code.