Widen the type of Mono.fromFuture, improve javadoc
This commit widens the type of `Mono.fromFuture` to accept any class
which is both a `CompletionStage` and a `Future`.

In the case of `fromFuture(CompletableFuture)`, the original signature
is kept for binary compatibility but an overload with the intersection
type is also provided.

In the case of `Supplier` based overloads and of methods that were just
introduced in the current snapshot, CompletableFuture type is replaced
with the intersection type.

Javadocs of all these methods as well as the fromCompletionStage methods
has been revised to better reflect this and hint at the cancellation
behavior.