Add .singleOptional() to Mono (#3317)
Introducing a new operator for the `Mono` stack.

It converts source `Mono` signals of type `T` to `Optional<T>`.

In case of a value, it is wrapped. When the source completes without a value, an
empty `Optional` is delivered. Errors are simply propagated downstream.

Having a dedicated operator for this case improves performance greatly over a
combination of operators:
`.map(Optional::ofNullable).defaultIfEmpty(Optional.empty())`.