Apply dependencies.gradle in root project
Dependabot seems to have trouble resolving the
dependencies.gradle file when it is used as a script plugin
in the subprojects section. However, it can be run in the
root build file, as it modifies the `ext` section,
which stands for [Extra properties](https://docs.gradle.org/current/userguide/writing_build_scripts.html#sec:extra_properties)
in Gradle's vocabulary and is made available to all the subprojects.