Remove Fuseable interface from ContextWriteRestoringThreadLocals (#3409)
`FluxContextWriteRestoringThreadLocals` and
`MonoContextWriteRestoringThreadLocals` implement `Fuseable`, but their
`Subscription` implementations don't support fusion.  The `FluxContextWrite`
and `MonoContextWrite` operators support fusion, but are not a hard `Thread`
barrier, such as the propagating counterparts. Fusion is forbidden in the
automatic propagation case as `ThreadLocal`s need to be set for every signal.
This simplification is compatible, as the modified classes are not public.