Propagate ThreadLocals in tap (#3405)
Aside from `contextWrite()`, the `tap` operator also has write access to the
`Context`. New classes have been added to handle the automatic context
propagation:

* `MonoTapRestoringThreadLocals`,
* `FluxTapRestoringThreadLocals`.

They replace the corresponding implementations of `tap` when
`Hooks.enableAutomaticContextPropagation()` is called, and when the
`context-propagation` library is on the class path.

As a result of these changes, `FluxTapTest` was also moved to the
`withMicrometerTest` source set, where the `context-propagation` library is
present.

Fixes #3395.