Transparent contextCapture in block operators (#3420)
Operators that block to retrieve the result from the reactive chain are usually
used in imperative contexts. If `ThreadLocal` values are present at call site,
they can be captured automatically. This change frees the user from calling
`contextCapture` everywhere that they block.

The following operators were altered to capture automatically if the
context-propagation library is on the classpath:

* `Mono#block`,
* `Mono#blockOptional`,
* `Flux#blockFirst`,
* `Flux#blockLast`,
* `Flux#toIterable`,
* `Flux#toStream`.